[Up] Convert everything to float : simplify

+++ src/coordinates.hpp
@@ -3,11 +3,11 @@
 
 class Coordinates {
 public:
-	int x, y;
+	float x, y;
 };
 
 bool operator<(const Coordinates& c1, const Coordinates& c2);
 bool operator>(const Coordinates& c1, const Coordinates& c2);
 
-std::array<int, 2> operator-(const Coordinates& c1, const Coordinates& c2);
-std::array<int, 2> operator+(const Coordinates& c1, const Coordinates& c2);
+std::array<float, 2> operator-(const Coordinates& c1, const Coordinates& c2);
+std::array<float, 2> operator+(const Coordinates& c1, const Coordinates& c2);
